src/main/java/za/ac/cput/service/ReviewService.java
package za.ac.cput.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import za.ac.cput.domain.Review;
import za.ac.cput.repository.ReviewRepository;

import java.util.List;
import java.util.Optional;

@Service
public class ReviewService {

    @Autowired
    private ReviewRepository reviewRepository;

    public Review create(Review review) {
        return reviewRepository.save(review);
    }

    public Optional<Review> read(long reviewId) {
        return reviewRepository.findById(reviewId);
    }

    public Review update(Review review) {
        return reviewRepository.save(review);
    }

    public boolean delete(long reviewId) {
        if (reviewRepository.existsById(reviewId)) {
            reviewRepository.deleteById(reviewId);
            return true;
        }
        return false;
    }

    public List<Review> getAll() {
        return reviewRepository.findAll();
    }


}
